+++ functionsystem/src/common/schedule_plugin/scorer/label_affinity_scorer/label_affinity_scorer.h
@@ -33,7 +33,7 @@ public:
                                         const resource_view::InstanceInfo &instance,
                                         const resource_view::ResourceUnit &resourceUnit) override;
 private:
-    int64_t CalculatePreferredScore(
+    double CalculatePreferredScore(
             const resource_view::ResourceUnit &resourceUnit, const resource_view::InstanceInfo &instance,
             const std::shared_ptr<schedule_framework::PreAllocatedContext> &preContext) const;
     bool isRelaxed_;
